Fixed login process

+++ OutbackWeatherTrackerApplication/src/main/java/cab302softwaredevelopment/outbackweathertrackerapplication/controllers/pages/SettingsController.java
@@ -73,6 +73,10 @@ public class SettingsController extends BasePage {
         } else if (event.getSource() == btnDeleteLocation) {
             Location selectedLocation = lstLocations.getSelectionModel().getSelectedItem();
             if (selectedLocation != null) {
+                if (lstLocations.getItems().size() == 1) {
+                    MainController.showAlert("Must have location", "Please add another location before deleting all locations");
+                    return;
+                }
                 locationService.deleteLocation(selectedLocation);
                 MainController.getController().updateUIData();
                 refreshLocationList();
